+++ cli/cli/commands/enclave/dump/dump.go
@@ -2,53 +2,19 @@ package dump
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/client"
-	"github.com/docker/docker/pkg/stdcopy"
-	"github.com/gammazero/workerpool"
 	"github.com/kurtosis-tech/container-engine-lib/lib"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/defaults"
-	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_manager"
 	"github.com/kurtosis-tech/kurtosis-cli/commons/positional_arg_parser"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/forever_constants"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"google.golang.org/protobuf/types/known/emptypb"
-	"io"
-	"io/ioutil"
-	"os"
-	"path"
 	"strings"
 )
 
 const (
 	enclaveIdArg     = "enclave-id"
 	outputDirpathArg = "output-dirpath"
-
-	shouldGetStoppedContainers = true
-
-	containerInspectResultFilename = "spec.json"
-	containerLogsFilename          = "output.log"
-
-	numContainersToDumpAtOnce = 20
-
-	// Permisssions for the files & directories we create as a result of the dump
-	createdDirPerms  = 0755
-	createdFilePerms = 0644
-
-	shouldCaptureContainerStderr = true
-	shouldCaptureContainerStdout = true
-	shouldFollowContainerLogs    = false
-
-	containerSpecJsonSerializationIndent = "  "
-	containerSpecJsonSerializationPrefix = ""
 )
 
 var positionalArgs = []string{
@@ -76,231 +42,11 @@ func run(cmd *cobra.Command, args []string) error {
 	enclaveId := parsedPositionalArgs[enclaveIdArg]
 	enclaveOutputDirpath := parsedPositionalArgs[outputDirpathArg]
 
-	// TODO REMOVE THIS ONCE EVERYTHING FLOWS THROUGH THE KURTOSISBACKEND!!
-	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating the Docker client")
-	}
-	dockerManager := docker_manager.NewDockerManager(
-		dockerClient,
-	)
-
 	kurtosisBackend, err := lib.GetLocalDockerKurtosisBackend()
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting a Kurtosis backend connected to local Docker")
-	}
-	engineManager := engine_manager.NewEngineManager(kurtosisBackend)
-
-	engineClient, closeClientFunc, err := engineManager.StartEngineIdempotentlyWithDefaultVersion(ctx, defaults.DefaultEngineLogLevel)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating a new Kurtosis engine client")
-	}
-	defer closeClientFunc()
-
-	getEnclavesResp, err := engineClient.GetEnclaves(ctx, &emptypb.Empty{})
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting enclaves, which is necessary to display the state for enclave '%v'", enclaveId)
-	}
-
-	if _, found := getEnclavesResp.EnclaveInfo[enclaveId]; !found {
-		return stacktrace.NewError("No enclave with ID '%v' exists", enclaveId)
-	}
-
-	// TODO REPLACE THIS WITH A CALL TO THE ENGINE SERVER TO GET THE CONTAINER IDS!!!
-	enclaveContainerSearchLabels := map[string]string{
-		forever_constants.AppIDLabel:   forever_constants.AppIDValue,
-		schema.EnclaveIDContainerLabel: enclaveId,
-	}
-
-	enclaveContainers, err := dockerManager.GetContainersByLabels(ctx, enclaveContainerSearchLabels, shouldGetStoppedContainers)
-	if err != nil {
-		return stacktrace.Propagate(
-			err,
-			"An error occurred getting the containers in enclave '%v' so their logs could be dumped to disk",
-			enclaveId,
-		)
-	}
-
-	// Create output directory
-	if _, err := os.Stat(enclaveOutputDirpath); !os.IsNotExist(err) {
-		return stacktrace.NewError("Cannot create output directory at '%v'; directory already exists", enclaveOutputDirpath)
-	}
-	if err := os.Mkdir(enclaveOutputDirpath, createdDirPerms); err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating output directory at '%v'", enclaveOutputDirpath)
-	}
-
-	workerPool := workerpool.New(numContainersToDumpAtOnce)
-	resultErrsChan := make(chan error, len(enclaveContainers))
-	for _, container := range enclaveContainers {
-		containerName := container.GetName()
-		containerId := container.GetId()
-		logrus.Debugf("Submitting job to dump info about container with name '%v' and ID '%v'", containerName, containerId)
-
-		/*
-			!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-			It's VERY important that the actual `func()` job function get created inside a helper function!!
-			This is because variables declared inside for-loops are created BY REFERENCE rather than by-value, which
-				means that if we inline the `func() {....}` creation here then all the job functions would get a REFERENCE to
-				any variables they'd use.
-			This means that by the time the job functions were run in the worker pool (long after the for-loop finished)
-				then all the job functions would be using a reference from the last iteration of the for-loop.
-
-			For more info, see the "Variables declared in for loops are passed by reference" section of:
-				https://www.calhoun.io/gotchas-and-common-mistakes-with-closures-in-go/ for more details
-			!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-		*/
-		jobToSubmit := createDumpContainerJob(
-			ctx,
-			dockerClient,
-			enclaveOutputDirpath,
-			resultErrsChan,
-			containerName,
-			containerId,
-		)
-
-		workerPool.Submit(jobToSubmit)
-	}
-	workerPool.StopWait()
-	close(resultErrsChan)
-
-	allResultErrStrs := []string{}
-	for resultErr := range resultErrsChan {
-		allResultErrStrs = append(allResultErrStrs, resultErr.Error())
-	}
-
-	if len(allResultErrStrs) > 0 {
-		allIndexedResultErrStrs := []string{}
-		for idx, resultErrStr := range allResultErrStrs {
-			indexedResultErrStr := fmt.Sprintf(">>>>>>>>>>>>>>>>> ERROR %v <<<<<<<<<<<<<<<<<\n%v", idx, resultErrStr)
-			allIndexedResultErrStrs = append(allIndexedResultErrStrs, indexedResultErrStr)
-		}
-
-		// NOTE: We don't use stacktrace here because the actual stacktraces we care about are the ones from the threads!
-		return errors.New(fmt.Sprintf(
-			"The following errors occurred when trying to dump information about enclave '%v':\n%v",
-			enclaveId,
-			strings.Join(allIndexedResultErrStrs, "\n\n"),
-		))
+	if err := kurtosisBackend.DumpEnclave(ctx, enclave.EnclaveID(enclaveId), enclaveOutputDirpath); err != nil {
+		return stacktrace.Propagate(err, "An error occurred dumping enclave '%v' to '%v'", enclaveId, enclaveOutputDirpath)
 	}
 
 	logrus.Infof("Dumped enclave '%v' to directory '%v'", enclaveId, enclaveOutputDirpath)
-	return nil
-}
-
-func createDumpContainerJob(
-	ctx context.Context,
-	dockerClient *client.Client,
-	enclaveOutputDirpath string,
-	resultErrsChan chan error,
-	containerName string,
-	containerId string,
-) func() {
-	return func() {
-		if err := dumpContainerInfo(ctx, dockerClient, enclaveOutputDirpath, containerName, containerId); err != nil {
-			resultErrsChan <- stacktrace.Propagate(
-				err,
-				"An error occurred dumping container info for container with name '%v' and ID '%v'",
-				containerName,
-				containerId,
-			)
-		}
-	}
-}
-
-func dumpContainerInfo(
-	ctx context.Context,
-	dockerClient *client.Client,
-	enclaveOutputDirpath string,
-	containerName string,
-	containerId string,
-) error {
-	// Make output directory
-	containerOutputDirpath := path.Join(enclaveOutputDirpath, containerName)
-	if err := os.Mkdir(containerOutputDirpath, createdDirPerms); err != nil {
-		return stacktrace.Propagate(
-			err,
-			"An error occurred creating directory '%v' to hold the output of container with name '%v' and ID '%v'",
-			containerOutputDirpath,
-			containerName,
-			containerId,
-		)
-	}
-
-	// Write container inspect results to file
-	inspectResult, err := dockerClient.ContainerInspect(
-		ctx,
-		containerId,
-	)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred inspecting container with ID '%v'", containerId)
-	}
-	jsonSerializedInspectResultBytes, err := json.MarshalIndent(inspectResult, containerSpecJsonSerializationPrefix, containerSpecJsonSerializationIndent)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred serializing the results of inspecting container with ID '%v' to JSON", containerId)
-	}
-	specOutputFilepath := path.Join(containerOutputDirpath, containerInspectResultFilename)
-	ioutil.WriteFile(specOutputFilepath, jsonSerializedInspectResultBytes, createdFilePerms)
-	if err != nil {
-		return stacktrace.Propagate(
-			err,
-			"An error occurred creating file '%v' to hold the spec of container with name '%v' and ID '%v'",
-			specOutputFilepath,
-			containerName,
-			containerId,
-		)
-	}
-
-	// Write container logs to file
-	containerLogsReadCloser, err := dockerClient.ContainerLogs(
-		ctx,
-		containerId,
-		types.ContainerLogsOptions{
-			ShowStderr: shouldCaptureContainerStderr,
-			ShowStdout: shouldCaptureContainerStdout,
-			Follow:     shouldFollowContainerLogs,
-		},
-	)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting the logs for container with ID '%v'", containerId)
-	}
-	defer containerLogsReadCloser.Close()
-	logsOutputFilepath := path.Join(containerOutputDirpath, containerLogsFilename)
-	logsOutputFp, err := os.Create(logsOutputFilepath)
-	if err != nil {
-		return stacktrace.Propagate(
-			err,
-			"An error occurred creating file '%v' to hold the logs of container with name '%v' and ID '%v'",
-			logsOutputFilepath,
-			containerName,
-			containerId,
-		)
-	}
-
-	// TODO Figure out a way to abstract this!!! This check-if-the-container-is-TTY-and-use-io.Copy-if-so-and-stdcopy-if-not
-	//  is copied straight from the Docker CLI, but it REALLY sucks that a Kurtosis dev magically needs to know that that's what
-	//  they have to do if they want to read container logs
-	// If we don't have this, reading the logs from REPL container breaks
-	if inspectResult.Config.Tty {
-		if _, err := io.Copy(logsOutputFp, containerLogsReadCloser); err != nil {
-			return stacktrace.Propagate(
-				err,
-				"An error occurred copying the TTY container logs stream to file '%v' for container with name '%v' and ID '%v'",
-				logsOutputFilepath,
-				containerName,
-				containerId,
-			)
-		}
-	} else {
-		if _, err := stdcopy.StdCopy(logsOutputFp, logsOutputFp, containerLogsReadCloser); err != nil {
-			return stacktrace.Propagate(
-				err,
-				"An error occurred copying the non-TTY container logs stream to file '%v' for container with name '%v' and ID '%v'",
-				logsOutputFilepath,
-				containerName,
-				containerId,
-			)
-		}
-	}
-
 	return nil
 }